# frontend/Agricultural_Products_Market_admin/default_OpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/order": {
      "get": {
        "tags": [
          "订单管理"
        ],
        "summary": "查看订单列表",
        "description": "获取当前用户的所有订单",
        "operationId": "getOrderList",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListOrderInfo"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "订单管理"
        ],
        "summary": "创建订单",
        "description": "根据购物车内容创建新订单",
        "operationId": "createOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultOrderInfo"
                }
              }
            }
          }
        }
      }
    },
    "/order/pay": {
      "post": {
        "tags": [
          "订单管理"
        ],
        "summary": "支付订单",
        "description": "模拟支付订单",
        "operationId": "payOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayOrderDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "认证管理"
        ],
        "summary": "用户注册",
        "description": "创建新用户账号",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "认证管理"
        ],
        "summary": "用户登录",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/auth/login": {
      "post": {
        "tags": [
          "管理员认证"
        ],
        "summary": "管理员登录",
        "description": "管理员登录接口，返回JWT令牌",
        "operationId": "login_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
...
    "/order/{id}": {
      "get": {
        "tags": [
          "订单管理"
        ],
        "summary": "查看订单详情",
        "description": "获取指定订单的详细信息，包括订单项",
        "operationId": "getOrderDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultOrderVO"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/order": {
      "get": {
        "tags": [
          "管理员订单管理"
        ],
        "summary": "分页展示所有用户订单",
        "description": "分页查询所有用户订单，支持按下单日期和支付状态筛选",
        "operationId": "getOrderList_1",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "payStatus",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultIPageOrderInfo"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResultString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "type": "string"
          }
        }
      },
...
      "CreateOrderDTO": {
        "type": "object",
        "properties": {
          "addressId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "OrderInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "订单ID",
            "format": "int64"
          },
          "orderNo": {
            "type": "string",
            "description": "订单编号"
          },
          "userId": {
            "type": "integer",
            "description": "用户ID",
            "format": "int64"
          },
          "addressId": {
            "type": "integer",
            "description": "地址ID",
            "format": "int64"
          },
          "totalAmount": {
            "type": "number",
            "description": "订单总金额"
          },
          "payStatus": {
            "type": "integer",
            "description": "支付状态：0-未支付，1-已支付",
            "format": "int32"
          },
          "orderStatus": {
            "type": "string",
            "description": "订单状态：待发货/已发货/已完成"
          },
          "payTime": {
            "type": "string",
            "description": "支付时间",
            "format": "date-time"
          },
          "createTime": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "description": "更新时间",
            "format": "date-time"
          }
        },
        "description": "订单表"
      },
      "ResultOrderInfo": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/OrderInfo"
          }
        }
      },
      "PayOrderDTO": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "RegisterDTO": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "LoginDTO": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "Result": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      },
...
      "ResultListOrderInfo": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderInfo"
            }
          }
        }
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "订单项ID",
            "format": "int64"
          },
          "orderId": {
            "type": "integer",
            "description": "订单ID",
            "format": "int64"
          },
          "produceId": {
            "type": "integer",
            "description": "农产品ID",
            "format": "int64"
          },
          "produceName": {
            "type": "string",
            "description": "农产品名称"
          },
          "price": {
            "type": "number",
            "description": "购买单价"
          },
          "quantity": {
            "type": "integer",
            "description": "购买数量",
            "format": "int32"
          },
          "createTime": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          }
        },
        "description": "订单项明细表"
      },
      "OrderVO": {
        "type": "object",
        "properties": {
          "orderInfo": {
            "$ref": "#/components/schemas/OrderInfo"
          },
          "orderItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          },
          "recipient": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          }
        }
      },
      "ResultOrderVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/OrderVO"
          }
        }
      },
...
      "IPageOrderInfo": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "pages": {
            "type": "integer",
            "format": "int64",
            "deprecated": true
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderInfo"
            }
          },
          "current": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ResultIPageOrderInfo": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/IPageOrderInfo"
          }
        }
      },
...
    }
  },
...
}